refactor: move Kaggle job utilities to a new backend module and update references in CLI and tests
